+++ tests/integration/fixtures/processing.py
@@ -6,6 +6,7 @@ import pytest
 import os
 import confluent_kafka as kafka
 from copy import deepcopy
+import json
 
 
 @pytest.fixture
@@ -82,6 +83,33 @@ def relay_with_processing(relay, mini_sentry, processing_config):
     return inner
 
 
+def kafka_producer(options):
+    # look for the servers (it is the only config we are interested in)
+    servers = [
+        elm["value"]
+        for elm in options["processing"]["kafka_config"]
+        if elm["name"] == "bootstrap.servers"
+    ]
+    if len(servers) < 1:
+        raise ValueError(
+            "Bad kafka_config, could not find 'bootstrap.servers'.\n"
+            "The configuration should have an entry of the format \n"
+            "{name:'bootstrap.servers', value:'127.0.0.1'} at path 'processing.kafka_config'"
+        )
+
+    servers = servers[0]
+
+    settings = {
+        "bootstrap.servers": servers,
+        "group.id": "test-consumer-%s" % uuid.uuid4().hex,
+        "enable.auto.commit": True,
+        "auto.offset.reset": "earliest",
+    }
+    producer = kafka.Producer(settings)
+
+    return producer
+
+
 @pytest.fixture
 def kafka_consumer(request, get_topic_name, processing_config):
     """
@@ -89,7 +117,8 @@ def kafka_consumer(request, get_topic_name, processing_config):
     """
 
     def inner(topic: str, options=None):
-        topics = [get_topic_name(topic)]
+        topic_name = get_topic_name(topic)
+        topics = [topic_name]
         options = processing_config(options)
         # look for the servers (it is the only config we are interested in)
         servers = [
@@ -120,24 +149,38 @@ def kafka_consumer(request, get_topic_name, processing_config):
             consumer.close()
 
         request.addfinalizer(die)
-        return consumer
+        return consumer, options, topic_name
 
     return inner
 
 
 class ConsumerBase(object):
-    def __init__(self, consumer):
+    def __init__(self, consumer, options, topic_name):
         self.consumer = consumer
+        self.test_producer = kafka_producer(options)
+        self.topic_name = topic_name
 
     # First poll takes forever, the next ones are fast
     def poll(self):
         rv = self.consumer.poll(timeout=5)
         return rv
 
+    def produce_test_message(self, message):
+        """
+        An associated producer, that can send message on the same topic as the
+        consumer used for tests when we don't expect anything to come back we
+        can send a test message at the end and verify that it is the first and
+        only message on the queue (care must be taken to make sure that the
+        test message ends up in the same partition as the message we are checking).
+        """
+        message_str = json.dumps(message)
+        self.test_producer.produce(self.topic_name, message_str)
+        self.test_producer.flush()
+
 
 @pytest.fixture
 def outcomes_consumer(kafka_consumer):
-    return lambda: OutcomesConsumer(kafka_consumer("outcomes"))
+    return lambda: OutcomesConsumer(*kafka_consumer("outcomes"))
 
 
 class OutcomesConsumer(ConsumerBase):
@@ -167,22 +210,22 @@ class OutcomesConsumer(ConsumerBase):
 
 @pytest.fixture
 def events_consumer(kafka_consumer):
-    return lambda: EventsConsumer(kafka_consumer("events"))
+    return lambda: EventsConsumer(*kafka_consumer("events"))
 
 
 @pytest.fixture
 def transactions_consumer(kafka_consumer):
-    return lambda: EventsConsumer(kafka_consumer("transactions"))
+    return lambda: EventsConsumer(*kafka_consumer("transactions"))
 
 
 @pytest.fixture
 def attachments_consumer(kafka_consumer):
-    return lambda: AttachmentsConsumer(kafka_consumer("attachments"))
+    return lambda: AttachmentsConsumer(*kafka_consumer("attachments"))
 
 
 @pytest.fixture
 def sessions_consumer(kafka_consumer):
-    return lambda: SessionsConsumer(kafka_consumer("sessions"))
+    return lambda: SessionsConsumer(*kafka_consumer("sessions"))
 
 
 class SessionsConsumer(ConsumerBase):